+++ image.cc
@@ -18,7 +18,7 @@ unsigned int Image::get_height() {
     return m_height;
 }
 
-void Image::colour_pixel(glm::dvec3 colour, unsigned int x, unsigned int y) {
+void Image::colour_pixel(glm::vec3 colour, unsigned int x, unsigned int y) {
     m_pixels[y * m_width + x] = colour; 
 }
